+++ apollo-biz/src/test/java/com/ctrip/framework/apollo/biz/config/BizConfigTest.java
@@ -24,7 +24,10 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.core.env.ConfigurableEnvironment;
+import org.springframework.core.env.Environment;
 import org.springframework.test.util.ReflectionTestUtils;
+
+import javax.sql.DataSource;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -42,11 +45,14 @@ public class BizConfigTest {
   @Mock
   private ServerConfigRepository serverConfigRepository;
 
+  @Mock
+  private DataSource dataSource;
+
   private BizConfig bizConfig;
 
   @Before
   public void setUp() throws Exception {
-    bizConfig = new BizConfig(new BizDBPropertySource(serverConfigRepository));
+    bizConfig = new BizConfig(new BizDBPropertySource(serverConfigRepository, dataSource, environment));
     ReflectionTestUtils.setField(bizConfig, "environment", environment);
   }