+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/portal/service/PortalDBPropertySource.java
@@ -25,9 +25,16 @@ import com.ctrip.framework.apollo.portal.repository.ServerConfigRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.env.Environment;
+import org.springframework.core.env.Profiles;
+import org.springframework.core.io.ClassPathResource;
+import org.springframework.core.io.Resource;
+import org.springframework.jdbc.datasource.init.DatabasePopulatorUtils;
+import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator;
 import org.springframework.stereotype.Component;
 
-import java.util.Map;
+import javax.annotation.PostConstruct;
+import javax.sql.DataSource;
 import java.util.Objects;
 
 
@@ -40,16 +47,27 @@ public class PortalDBPropertySource extends RefreshablePropertySource {
 
   private final ServerConfigRepository serverConfigRepository;
 
-  public PortalDBPropertySource(final String name,
-      final Map<String, Object> source,
-      final ServerConfigRepository serverConfigRepository) {
-    super(name, source);
-    this.serverConfigRepository = serverConfigRepository;
-  }
+  private final DataSource dataSource;
+
+  private final Environment env;
+
   @Autowired
-  public PortalDBPropertySource(final ServerConfigRepository serverConfigRepository) {
+  public PortalDBPropertySource(final ServerConfigRepository serverConfigRepository, DataSource dataSource,
+                                final Environment env) {
     super("DBConfig", Maps.newConcurrentMap());
     this.serverConfigRepository = serverConfigRepository;
+    this.dataSource = dataSource;
+    this.env = env;
+  }
+
+  @PostConstruct
+  public void runSqlScript() throws Exception {
+    if (env.acceptsProfiles(Profiles.of("h2"))) {
+      Resource resource = new ClassPathResource("jpa/init.h2.sql");
+      if (resource.exists()) {
+        DatabasePopulatorUtils.execute(new ResourceDatabasePopulator(resource), dataSource);
+      }
+    }
   }
 
   @Override